fix: Export Makefile variables to child processes
The issue was that when Makefile included .env, the variables were not
exported to child processes. When utils.sh was executed via
'make verify-files', it couldn't access MANIFESTS_DIR from .env, causing
HELM_CHARTS_DIR to be set to '/helm-charts-values' instead of
'manifests/helm-charts-values'.

This resulted in the error: [ERROR] /ovn-values.yaml not found

Fix: Added 'export' directive after 'include .env' in Makefile to ensure
all variables are available to child processes.